исправлено все, кроме переноса функций из App в форму
