+++ src/components/article-params-form/ArticleParamsForm.module.scss
@@ -22,7 +22,7 @@
 	box-sizing: border-box;
 
 	// подкорректировал ширину формы, чтобы убрать горизнтальный скорлл
-	width: 591px;
+	width: 616px;
 	height: auto;
 	min-height: 100%;
 	padding: 50px 31px;
